Former-commit-id: c2948ffb7816c3c785fc512cc3286b4ad2eb9040

+++ rec-common/src/main/java/es/capgemini/pfs/zona/dao/ZonaDao.java
@@ -78,6 +78,6 @@ public interface ZonaDao extends AbstractDao<DDZona, Long> {
      */
     Boolean existePerfilZona(Long idZona, Long idPerfil);
     
-    List<DDZona> getZonasJerarquiaByCodDesc(Long idNivel, Set<String> codigoZonasUsuario, String codDesc);
+    List<DDZona> getZonasJerarquiaByCodDesc(Integer idNivel, Set<String> codigoZonasUsuario, String codDesc);
 
 }
